+++ src/main/scala/com/github/agourlay/cornichon/resolver/Resolvable.scala
@@ -1,0 +1,16 @@
+package com.github.agourlay.cornichon.http
+
+trait Resolvable[A] {
+
+  def toResolvableForm(r: A): String
+  def fromResolvableForm(r: String): A
+}
+
+object Resolvable {
+
+  implicit val stringResolvableForm = new Resolvable[String] {
+    def toResolvableForm(s: String) = s
+    def fromResolvableForm(s: String) = s
+  }
+
+}
